Home redesign, account view, Basic pan tool, Freeform overlay + vibrance
Home page:
- Remove "created by Micah Forstein" byline; add "Chainmail Studio" subtitle.
- Rename release-notes section to "Documentation"; drop dynamic note content.
- New "Plan" panel: Pricing, current plan (Free unless signed in with a Stripe
  customer), Sign In/Out. Move EULA/Privacy to buttons; remove the footer.
- Center the Navigator/Choose-a-workspace/Utilities/Documentation headings.
- Remove the signed-in email from the Navigator.

Auth: /auth now shows an account view (Signed in as {email} + Sign Out) instead
of bouncing signed-in users away.

Release notes: data + numeric date now read "Micah, 6/10/2026".

Basic: add a dedicated ✋ Pan button (paint/pan toggle pair).

Freeform: rings follow the Tuner tilt on load + restore Ring Angle sliders;
image transfer keeps full color and ignores opacity (no white fade); narrower
overlay panel with a fixed header (✕ always visible); Transfer button moved
above the preview.

RingRenderer: image-filled scales render vibrant (sRGB texture + emissive lift).

+++ src/pages/ReleaseNotes.tsx
@@ -100,7 +100,7 @@ export default function ReleaseNotes() {
               {e.content}
             </div>
             <div style={{ color: "#6b7280", fontSize: 12 }}>
-              — {e.author}, {new Date(e.timestamp).toLocaleDateString("en-US", { year: "numeric", month: "long", day: "numeric" })}
+              — {e.author}, {new Date(e.timestamp).toLocaleDateString("en-US", { year: "numeric", month: "numeric", day: "numeric" })}
             </div>
           </div>
         ))}
